% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/onto.r
\name{ontobrowse}
\alias{ontobrowse}
\title{Ontobrowse helps you browse and maintain your ontology}
\usage{
ontobrowse(onto = TRUE)
}
\arguments{
\item{onto}{Does your ontology exists? Defaults to TRUE.}
}
\description{
This function allows you browse.
}
\examples{
ontobrowse()
}
\keyword{ontology}

