% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/ontobrowse.r
\name{ontologies}
\alias{ontologies}
\title{Ontologies helps you browse your ontologies}
\usage{
ontologies()
}
\description{
This function allows you to see what ontologies are available.
no param
ontologies()
}

