% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/ontobrowse.r
\name{path}
\alias{path}
\title{Path shows the ancestry path from the root of the ontology to the given term ID. Does not support multiple inheritance now.}
\usage{
path(term, lang = "la", origin = 9000, returnIds = FALSE)
}
\arguments{
\item{term}{Give a term ID.}

\item{lang}{Give a language. Default: "la"}

\item{origin}{Give an origin. Default: 9000}

\item{returnIds}{Return IDs or terms in the provided language. Default: FALSE
path()}
}
\description{
This function outputs an array of IDs or of terms in the language provided. Order: term ID ->...-> origin
}

