% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/ontobrowse.r
\name{smp_api}
\alias{smp_api}
\title{Sapiens Mapping API allows accessing anatomy elements from the 2D atlas by id}
\usage{
smp_api(term)
}
\arguments{
\item{term}{Give a term ID
smp_api()}
}
\description{
This function outputs the SMP color used for the element and the minimum, maximum and intermediary values for the slices in which the element is present.
}

