% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/ontobrowse.r
\name{translations}
\alias{translations}
\title{Translations gives all the translations found in the ontologies, for a term ID}
\usage{
translations(term)
}
\arguments{
\item{term}{Give a term ID
translations()}
}
\description{
This function outputs an array with key = language (ex."la","en" etc.) and value = translation
}

