% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/ontobrowse.r
\name{ontobrowse}
\alias{ontobrowse}
\title{Ontobrowse helps you browse your ontology}
\usage{
ontobrowse(term = 9000, lang = "la", origin = 9000)
}
\arguments{
\item{term}{Give a term_id. Defaults to 9000 (Terminologia Morphologica).}

\item{lang}{Give a display language (la=Latin, en=English, ro=Romanian). Defaults to "la" (Latin).}

\item{origin}{Give a term_id for the origin. Defaults to 9000 (Terminologia Morphologica).}
}
\description{
This function allows you browse an ontology by id.
}
\examples{
ontobrowse()
}
\keyword{ontology}

